Reorg create appointment page
